+++ WelcomeHome/WelcomeHome.DAL/Repositories/IDocumentRepository.cs
@@ -6,11 +6,11 @@ public interface IDocumentRepository
 {
     IEnumerable<Document> GetAll();
 
-    Task<Document?> GetByIdAsync(int id);
+    Task<Document?> GetByIdAsync(long id);
 
     Task AddAsync(Document document);
 
     Task UpdateAsync(Document document);
 
-    Task DeleteAsync(int id);
+    Task DeleteAsync(long id);
 }
